refactor: migrate project_context_management
