+++ front/lib/api/actions/servers/project_context_management/index.ts
@@ -1,9 +1,10 @@
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
-import { registerFileTools } from "@app/lib/actions/mcp_internal_actions/servers/project_context_management/file_operations";
-import { registerMetadataTools } from "@app/lib/actions/mcp_internal_actions/servers/project_context_management/metadata_operations";
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
+import { registerTool } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
+import { PROJECT_CONTEXT_MANAGEMENT_SERVER_NAME } from "@app/lib/api/actions/servers/project_context_management/metadata";
+import { createProjectContextManagementTools } from "@app/lib/api/actions/servers/project_context_management/tools";
 import type { Authenticator } from "@app/lib/auth";
 
 /**
@@ -17,13 +18,14 @@ function createServer(
   auth: Authenticator,
   agentLoopContext?: AgentLoopContextType
 ): McpServer {
-  const server = makeInternalMCPServer("project_context_management");
+  const server = makeInternalMCPServer(PROJECT_CONTEXT_MANAGEMENT_SERVER_NAME);
 
-  // Register all file-related tools.
-  registerFileTools(server, auth, agentLoopContext);
-
-  // Register all metadata-related tools.
-  registerMetadataTools(server, auth, agentLoopContext);
+  const tools = createProjectContextManagementTools(auth, agentLoopContext);
+  for (const tool of tools) {
+    registerTool(auth, agentLoopContext, server, tool, {
+      monitoringName: PROJECT_CONTEXT_MANAGEMENT_SERVER_NAME,
+    });
+  }
 
   return server;
 }
