+++ front/lib/api/actions/servers/project_context_management/metadata.ts
@@ -1,0 +1,145 @@
+import type { JSONSchema7 as JSONSchema } from "json-schema";
+import { z } from "zod";
+import { zodToJsonSchema } from "zod-to-json-schema";
+
+import type { ServerMetadata } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+import { createToolsRecord } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+
+export const PROJECT_CONTEXT_MANAGEMENT_SERVER_NAME =
+  "project_context_management" as const;
+
+export const PROJECT_CONTEXT_MANAGEMENT_TOOLS_METADATA = createToolsRecord({
+  list_project_files: {
+    description:
+      "List all files in the project context. Returns file metadata including names, IDs, and content types.",
+    schema: {},
+    stake: "never_ask",
+  },
+  add_project_file: {
+    description:
+      "Add a new file to the project context. The file will be available to all conversations in this project. " +
+      "Provide either 'content' (text string) or 'sourceFileId' (ID of an existing file from the conversation to copy from).",
+    schema: {
+      fileName: z.string().describe("Name of the file to add"),
+      content: z
+        .string()
+        .optional()
+        .describe(
+          "Text content of the file (provide either this or sourceFileId)"
+        ),
+      sourceFileId: z
+        .string()
+        .optional()
+        .describe(
+          "ID of an existing file to copy content from (provide either this or content)"
+        ),
+      contentType: z
+        .string()
+        .optional()
+        .describe(
+          "MIME type (default: text/plain, or inherited from sourceFileId if provided)"
+        ),
+    },
+    stake: "high",
+  },
+  update_project_file: {
+    description:
+      "Update the content of an existing file in the project context. This replaces the entire file content. " +
+      "Provide either 'content' (text string) or 'sourceFileId' (ID of an existing file from the conversation to copy from).",
+    schema: {
+      fileId: z.string().describe("ID of the file to update"),
+      content: z
+        .string()
+        .optional()
+        .describe(
+          "New text content for the file (provide either this or sourceFileId)"
+        ),
+      sourceFileId: z
+        .string()
+        .optional()
+        .describe(
+          "ID of an existing file to copy content from (provide either this or content)"
+        ),
+    },
+    stake: "high",
+  },
+  edit_project_description: {
+    description:
+      "Edit the project description. This updates the project's description text.",
+    schema: {
+      description: z
+        .string()
+        .describe("New project description (free-form text)."),
+    },
+    stake: "low",
+  },
+  add_project_url: {
+    description:
+      "Add a new URL to the project. URLs are named links (e.g., documentation, repository, design files).",
+    schema: {
+      name: z
+        .string()
+        .describe("Name/label for the URL (e.g., 'Documentation')"),
+      url: z.string().describe("The URL to add"),
+    },
+    stake: "low",
+  },
+  edit_project_url: {
+    description:
+      "Edit an existing URL in the project. You can change the name and/or the URL itself. " +
+      "Identify the URL to edit by its current name.",
+    schema: {
+      currentName: z.string().describe("Current name/label of the URL to edit"),
+      newName: z
+        .string()
+        .optional()
+        .describe("New name/label for the URL (leave empty to keep current)"),
+      newUrl: z
+        .string()
+        .optional()
+        .describe("New URL value (leave empty to keep current)"),
+    },
+    stake: "low",
+  },
+  read_project_journal_entry: {
+    description:
+      "Reads all journal entries for this project. Returns the journal entries with their content, and timestamps.",
+    schema: {
+      limit: z
+        .number()
+        .optional()
+        .describe("Maximum number of journal entries to return (default: 1)"),
+    },
+    stake: "never_ask",
+  },
+});
+
+const PROJECT_CONTEXT_MANAGEMENT_INSTRUCTIONS =
+  "Project context files are shared across all conversations in this project. " +
+  "Only text-based files are supported for adding/updating. " +
+  "You can add/update files by providing text content directly, or by copying from existing files (like those you've generated). " +
+  "Requires write permissions on the project space.";
+
+export const PROJECT_CONTEXT_MANAGEMENT_SERVER = {
+  serverInfo: {
+    name: "project_context_management",
+    version: "1.0.0",
+    description:
+      "Manage files in the project context. Add, update, delete, and list project files.",
+    icon: "ActionDocumentTextIcon",
+    authorization: null,
+    documentationUrl: null,
+    instructions: PROJECT_CONTEXT_MANAGEMENT_INSTRUCTIONS,
+  },
+  tools: Object.values(PROJECT_CONTEXT_MANAGEMENT_TOOLS_METADATA).map((t) => ({
+    name: t.name,
+    description: t.description,
+    inputSchema: zodToJsonSchema(z.object(t.schema)) as JSONSchema,
+  })),
+  tools_stakes: Object.fromEntries(
+    Object.values(PROJECT_CONTEXT_MANAGEMENT_TOOLS_METADATA).map((t) => [
+      t.name,
+      t.stake,
+    ])
+  ),
+} as const satisfies ServerMetadata;
